bisect: add bisect next documentation
`git bisect next` is already mentioned in the help text from `git bisect
help`, but isn't actually covered in the bisect documentation, so add a
short paragraph on it.

Also remove the decade-old comment about whether "next" should be a
UI-level command: it is a useful command that doesn't have an obvious
alternative.

Signed-off-by: Adam Dinwoodie <adam@dinwoodie.org>